% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htk_CyC.R
\name{htk_CyC}
\alias{htk_CyC}
\title{compare-your-country plot for each selected indicator in the housing toolkit chapters}
\usage{
htk_CyC(mydata, ctry, var_codes, var_names, sec_col, title = NULL)
}
\arguments{
\item{mydata}{dataframe that contains the variables and dates to be chosen}

\item{ctry}{selected OECD countries for which the country note is produced}

\item{var_codes}{list of indicators}

\item{var_names}{names of the indicators}

\item{sec_col}{colours for each section}

\item{title}{a title for the figure}
}
\value{
returns a standardized plot
}
\description{
The functions generates a (line) plot displaying the value of the selected indicator for each country and with the average OECD value, the bottom and top performing countries.
}
\author{
Manuel Betin
Federica De Pace
}
