\name{htk_line_extract}
\alias{htk_line_extract}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract chosen variables per country
}
\description{
Extracts the data per country and relevant variable in a data frame
}
\usage{
htk_line_extract(data_source, country_code, var1,var2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data_source}{ data set used to extract the chosen country and variables}
  \item{country_code}{ISO3_code of the chosen country }
  \item{var1}{variable to be extracted}
  \item{var2}{variable to be extracted}
}
\author{
Maxime Nguyen
Manuel Betin
}
