% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htk_lineplot.R
\name{htk_lineplot}
\alias{htk_lineplot}
\title{Line barplots with OECD template style to display timeseries evolutions in the housing toolkit chapters}
\usage{
htk_lineplot(data = resilience_database, subtitle = ".",
  title = NULL, Xlabel = "Date", Ylabel = "Price index", ymin = 0,
  ymax = 120, ybreak = 10, xbreak = "5 years", start_date = 1970,
  end_date = 2018, file = NULL, path = NULL, width = 4,
  height = 4, dpi = 300)
}
\arguments{
\item{data}{dataframe to be used for the plot}

\item{subtitle}{subtitle of the plot}

\item{title}{title of the plot}

\item{Xlabel}{label of the X axis}

\item{Ylabel}{label of the Y axis}

\item{start_date}{data start date}

\item{end_date}{data end date}

\item{file}{name of the file containing the plot were it to be saved, name must be followed by .png}

\item{path}{path of where the file is to be saved}

\item{width}{width of the image to be saved}

\item{height}{height of the image to be saved}

\item{dpi}{resolution}

\item{Ymin}{minimum y value to be displayed}

\item{Ymax}{maximum y value to be displayed}

\item{Ybreak}{difference between Y ticks}

\item{Xbreak}{difference between X ticks}
}
\value{
Returns a standardized single line plot
}
\description{
Function that plots a lines in a standardized way. The function is built on the assumption that the data is structured in the agreed format (first column as countries, second as period, others as key variables)
}
\author{
Manuel Betin
Maxime Nguyen
}
