% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htk_policyradar.R
\name{htk_policyradar}
\alias{htk_policyradar}
\title{Radar chart for policy variables}
\usage{
htk_policyradar(data, ctry, title = NULL)
}
\arguments{
\item{data}{a dataframe with the policy variables in
columns and iso3 codes in rows.}

\item{ctry}{the iso3 code for the selected country}

\item{title}{an optional title for the figure}
}
\value{
fmsl radarchart object
}
\description{
Radar chart to produce the spider graph
to display the selected policy variables
}
\author{
Manuel Betin
}
