\name{resilience_database}
\alias{resilience_database}
\docType{data}
\title{
Dataset containing the variables to show for the resilience chapter
}
\description{
Dataset containing the variables to show for the resilience chapter
}
\usage{data("resilience_database")}
\format{
  A data frame with 10620 observations on the following 48 variables.
  \describe{
    \item{\code{ISO3_code}}{a character vector}
    \item{\code{Period}}{a character vector}
    \item{\code{gdp}}{a numeric vector}
    \item{\code{per}}{a numeric vector}
    \item{\code{dwelst}}{a numeric vector}
    \item{\code{LTVP_ext}}{a numeric vector}
    \item{\code{cr}}{a numeric vector}
    \item{\code{cr_tier1}}{a numeric vector}
    \item{\code{cr_mortgage}}{a numeric vector}
    \item{\code{hpi}}{a numeric vector}
    \item{\code{hpi_ydh}}{a numeric vector}
    \item{\code{ihq}}{a numeric vector}
    \item{\code{iph}}{a numeric vector}
    \item{\code{lbanfq}}{a numeric vector}
    \item{\code{lcdr}}{a numeric vector}
    \item{\code{lchq}}{a numeric vector}
    \item{\code{lchr}}{a numeric vector}
    \item{\code{lhor}}{a numeric vector}
    \item{\code{lmhq}}{a numeric vector}
    \item{\code{lmhr}}{a numeric vector}
    \item{\code{lthq}}{a numeric vector}
    \item{\code{rhp}}{a numeric vector}
    \item{\code{rpi}}{a numeric vector}
    \item{\code{tier1kar}}{a numeric vector}
    \item{\code{tier1r}}{a numeric vector}
    \item{\code{lghq}}{a numeric vector}
    \item{\code{lghr}}{a numeric vector}
    \item{\code{DIW_rent_laws}}{a numeric vector}
    \item{\code{DIW_tenure_security}}{a numeric vector}
    \item{\code{DIW_rationing}}{a numeric vector}
    \item{\code{Eurostat_Construction_Rate}}{a numeric vector}
    \item{\code{Owners}}{a numeric vector}
    \item{\code{social_rent}}{a numeric vector}
    \item{\code{spending_housing}}{a numeric vector}
    \item{\code{prop_tax_equity_own}}{a numeric vector}
    \item{\code{prop_tax_equity_rent}}{a numeric vector}
    \item{\code{credit_hh}}{a numeric vector}
    \item{\code{credit_gap_gdp}}{a numeric vector}
    \item{\code{credit_gdp}}{a numeric vector}
    \item{\code{max_hpi_rpi}}{a numeric vector}
    \item{\code{min_hpi_rpi}}{a numeric vector}
    \item{\code{last_hpi_rpi}}{a numeric vector}
    \item{\code{Owner_occupied}}{a numeric vector}
    \item{\code{Rented}}{a numeric vector}
    \item{\code{hpi_rpi}}{a numeric vector}
    \item{\code{price_to_income}}{a numeric vector}
    \item{\code{peak_to_trough_x}}{a numeric vector}
    \item{\code{time_to_recovery}}{a numeric vector}
  }
}
\keyword{datasets}
