% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htk_barplot.R
\name{htk_barplot}
\alias{htk_barplot}
\title{barplots with OECD template style to display cross country comparisons in the housing toolkit chapters}
\usage{
htk_barplot(data_source, x, y, y2 = NULL, y3 = NULL, Xlabel = NULL,
  Ylabel = NULL, ylegend = NULL, y2legend = NULL, y3legend = NULL,
  title = NULL, subtitle = NULL, file = NULL, path1 = NULL)
}
\arguments{
\item{data_source}{dataframe that contains the variables and dates to be chosen}

\item{x}{x axis variable, often OECD countries}

\item{y}{y axis variable to be plotted}

\item{y2}{other y variable to be plotted}

\item{y3}{other y variable to be plotted}

\item{Xlabel}{label of the X axis}

\item{Ylabel}{label of the Y axis}

\item{ylegend}{legend of the first y variable}

\item{y2legend}{legend of the second y variable}

\item{y3legend}{legend of the third y variable}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{file}{name of the file that contains the plot if you decide to save it, filename must end with .png}

\item{path1}{specify exit path of the file}
}
\value{
returns a standardized barplot
}
\description{
The functions generates barplots displaying the value of the selected indicator for each country and with the average OECD value in highlighted. The plot respect the OECD style and allow the generation of standardize figures for all the chapters of the housing toolkit
}
\author{
Manuel Betin
Maxime Nguyen
}
