% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htk_scatterplot.R
\name{htk_scatterplot}
\alias{htk_scatterplot}
\title{scatter barplots with OECD template style to display correlation between two variables in the housing toolkit chapters}
\usage{
htk_scatterplot(
  data_source,
  myvar_x,
  myvar_y,
  my_label,
  Xlabel = NULL,
  Ylabel = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{data_source}{dataframe that contains the variables to plot}

\item{myvar_x}{x axis variable name}

\item{myvar_y}{y axis variable name}

\item{my_label}{labels used to label the points in the plot}

\item{Xlabel}{label of the X axis}

\item{Ylabel}{label of the Y axis}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}
}
\value{
returns a standardized scatter plot
}
\description{
The functions generates scatter plots in an agreed format to ensure presentation consistency in the dashboard
}
\author{
Manuel Betin
Maxime Nguyen
}
