% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htk_text_generator.r
\name{htk_text_generator}
\alias{htk_text_generator}
\title{Create dynamique text from}
\usage{
htk_text_generator(data, category, vars)
}
\arguments{
\item{data}{dataset with one row per country}

\item{category}{name of the category: Efficiency, Affordability, Sustainability}

\item{vars}{a vector of four variable names that enter the algorithm}
}
\value{
a dataset containing the values of the indicator for each country
and the paragraph of text interpretating the data
}
\description{
generate dynamic text
}
\author{
Manuel Betin, Federica Depace, Naomi Cohen
}
